Revert "ServerInteraction"
