+++ package.json
@@ -18,7 +18,6 @@
     "@react-navigation/stack": "^6.4.1",
     "@tensorflow/tfjs": "^4.20.0",
     "@tensorflow/tfjs-react-native": "^1.0.0",
-    "axios": "^1.7.7",
     "cosmiconfig": "^9.0.0",
     "expo-asset": "~10.0.10",
     "expo-blur": "^13.0.2",
@@ -47,8 +46,8 @@
     "react-native-vector-icons": "^10.1.0",
     "react-native-video": "^6.6.3",
     "react-native-webview": "^13.12.2",
-    "react-native-worklets-core": "^1.3.3",
     "sqlite3": "^5.1.7"
+    "react-native-worklets-core": "^1.3.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
