+++ src/Components/LinkButton/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {Text} from 'react-native';
 import {getStyles} from './defaultCSS';
 import {Button} from 'react-native-paper';
-import {useSettingsContext} from '../../Contexts/FontContext';
+import {useSettingsContext} from '../../Contexts/SettingsContext';
 
 function Index(props) {
   const {selectedConfig, setSelectedConfig} = useSettingsContext();
