+++ src/Components/Text/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Text} from 'react-native-paper';
 import {getStyles} from './defaultCSS';
-import {useSettingsContext} from '../../Contexts/FontContext';
+import {useSettingsContext} from '../../Contexts/SettingsContext';
 
 function Index(props) {
   const {selectedConfig, setSelectedConfig} = useSettingsContext();
