+++ src/Screens/Home/index.tsx
@@ -21,44 +21,41 @@ import BottomNav from '../../Features/BottomNavigation';
 import Video from '../Video';
 import ColorblindSettings from '../ColorblindSettings';
 import CurriculumTodo from '../CurriculumTodo';
-import ServerInteraction from '../ServerInteraction';
 
 const Stack = createNativeStackNavigator();
 
 const Index = () => {
   return (
     <>
-      <AuthProvider>
-        <Stack.Navigator
-          initialRouteName=""
-          screenOptions={{
-            headerShown: false,
-          }}>
-          <Stack.Screen name="Welcome" component={WelcomeToSeal} />
+     <AuthProvider>
+      <Stack.Navigator
+        initialRouteName="WebCamFeed"
+        screenOptions={{
+          headerShown: false,
+        }}>
+        <Stack.Screen name="Welcome" component={WelcomeToSeal} />
+        <Stack.Screen
+          name="ColorblindSettings"
+          component={ColorblindSettings}
+        />
           <Stack.Screen
-            name="ColorblindSettings"
-            component={ColorblindSettings}
-          />
-          <Stack.Screen name="Curriculum Todo" component={CurriculumTodo} />
-          {/* <Stack.Screen name="Settings" component={Settings} /> */}
-          <Stack.Screen name="Login" component={Login} />
-          <Stack.Screen name="Personal" component={PersonalPage} />
-          <Stack.Screen name="AccountSignUp" component={AccountSignUp} />
-          {/* the "Main "screen below renders the bottom Nav which wraps the relevant pages it is included on */}
-          <Stack.Screen name="Main" component={BottomNav} />
-          <Stack.Screen name="HowSealWorks" component={HowSealWorks} />
-          {/* Commented for testing, uncomment to enable break screen*/}
-          <Stack.Screen name="Break" component={Break} />
-          <Stack.Screen name="Video" component={Video} />
-          {/* Commented for testing, to enable camera feed uncomment this */}
-          {/* <Stack.Screen name="WebCamFeed" component={WebCamFeed}  /> */}
-
-          <Stack.Screen
-            name="ServerInteraction"
-            component={ServerInteraction}
-          />
-        </Stack.Navigator>
-        <ColorblindFilter />
+          name="Curriculum Todo"
+          component={CurriculumTodo}
+        />
+        {/* <Stack.Screen name="Settings" component={Settings} /> */}
+        <Stack.Screen name="Login" component={Login} />
+        <Stack.Screen name="Personal" component={PersonalPage} />
+        <Stack.Screen name="AccountSignUp" component={AccountSignUp} />
+        {/* the "Main "screen below renders the bottom Nav which wraps the relevant pages it is included on */}
+        <Stack.Screen name="Main" component={BottomNav} />
+        <Stack.Screen name="HowSealWorks" component={HowSealWorks} />
+        {/* Commented for testing, uncomment to enable break screen*/}
+        <Stack.Screen name="Break" component={Break} />
+        <Stack.Screen name='Video' component={Video}/>
+        {/* Commented for testing, to enable camera feed uncomment this */}
+        {/* <Stack.Screen name="WebCamFeed" component={WebCamFeed}  /> */}
+      </Stack.Navigator>
+      <ColorblindFilter />
       </AuthProvider>
     </>
   );
