+++ src/Screens/PersonalPage/index.tsx
@@ -1,29 +1,39 @@
-import {Image, View} from 'react-native';
+import React from 'react';
+import {Dimensions, Image, TouchableOpacity, View} from 'react-native';
 import BackgroundWrapper from '../../Components/BackgroundWrapper';
 import Button from '../../Components/Button';
 import Text from '../../Components/Text';
 import {useAuth} from '../../Components/Authentication/AuthProvider';
 import ProtectedRoute from '../../Components/Authorization/ProtectedRoute';
 import {getStyles} from './defaultCSS';
-import {useSettingsContext} from '../../Contexts/FontContext';
+import {useSettingsContext} from '../../Contexts/SettingsContext';
 import PracticeSession from '../../Assets/svg/practice_session.svg';
 import AssignTasks from '../../Assets/svg/assign_tasks.svg';
-import {Appbar} from 'react-native-paper';
+import {Appbar, Icon} from 'react-native-paper';
+import {NavigationProp} from '@react-navigation/native';
+import CurriculumInput from '../CurriculumInput';
 
-const Index = ({navigation}) => {
+const Index = ({navigation}: {navigation: NavigationProp<any>}) => {
   const {selectedConfig} = useSettingsContext();
   const styles = getStyles(selectedConfig.font);
   const {currentUser} = useAuth();
+
+  const {width, height} = Dimensions.get('window');
 
   return (
     <BackgroundWrapper>
       <Appbar.BackAction onPress={() => navigation.goBack()} />
       <ProtectedRoute allowedRoles={['admin', 'viewer']}>
+        <TouchableOpacity
+          style={styles.settingsButton}
+          onPress={() => navigation.navigate('Settings')}>
+          <Icon size={30} source="cog" />
+        </TouchableOpacity>
         <View style={styles.container}>
           <View style={styles.section}>
             {/* <Text style={styles.header}>Welcome, {currentUser.name}</Text> */}
             <Text style={styles.subheader}>Activities</Text>
-            <View style={styles.row}>
+            <View style={width > 640 ? styles.row : styles.col}>
               <View style={styles.rowItem}>
                 <PracticeSession />
                 {/* <Text style={styles.itemTitle}>Practice sessions</Text> */}
@@ -55,20 +65,19 @@ const Index = ({navigation}) => {
               icon="trophy"
               onPress={() => navigation.navigate('Achievements')}
             />
-            <Button
-              title="Settings"
-              onPress={() => navigation.navigate('Settings')}
-            />
-            <Button
-              title="Message Teacher"
-              onPress={() => navigation.navigate('ServerInteraction')}
-            />
-            {/*Test button: Comment it to remove it. Will need to change Settings to 'NONE' before to remove filter*/}
-            <Button
-              title="Colorblind Settings"
-              onPress={() => navigation.navigate('ColorblindSettings')}
-            />
-          </View>
+
+          {/*Test button: Comment it to remove it. Will need to change Settings to 'NONE' before to remove filter*/}
+          <Button
+            title="Colorblind Settings"
+            onPress={() => navigation.navigate('ColorblindSettings')}
+          />
+
+          {/* Curriculum Component */}
+          <Button
+            title="Curriculum To Do"
+            icon="clipboard-list"
+            onPress={() => navigation.navigate('Curriculum Todo')}
+          />
         </View>
       </ProtectedRoute>
     </BackgroundWrapper>
