+++ src/Screens/Settings/index.tsx
@@ -11,7 +11,10 @@ import {
 
 import {getStyles} from './defaultCSS';
 import {useNotification} from '../../Features/useNotification';
-import {FontConfigType, useSettingsContext} from '../../Contexts/FontContext';
+import {
+  FontConfigType,
+  useSettingsContext,
+} from '../../Contexts/SettingsContext';
 import {NavigationProp} from '@react-navigation/native';
 import {Icon} from 'react-native-paper';
