+++ src/Screens/WelcomeToSeal/index.tsx
@@ -5,7 +5,7 @@ import {View, Text} from 'react-native';
 import BackgroundWrapper from '../../Components/BackgroundWrapper';
 import Button from '../../Components/Button';
 import getStyles from './defaultCSS';
-import {useSettingsContext} from '../../Contexts/FontContext';
+import {useSettingsContext} from '../../Contexts/SettingsContext';
 import Seal from '../../Assets/svg/seal.svg';
 import Hello from '../../Assets/svg/hello.svg';
 import SessionOptimizerComponent from '../../prediction/sessionPrediction';
@@ -17,6 +17,7 @@ const Index = ({navigation}: {navigation: NavigationProp<any>}) => {
 
   const {selectedConfig} = useSettingsContext();
   const styles = getStyles(selectedConfig.font);
+
 
   return (
     <BackgroundWrapper>
