+++ src/Styles/defaultCSS.tsx
@@ -1,5 +1,5 @@
 import {StyleSheet} from 'react-native';
-import {FontConfigType} from '../Contexts/FontContext';
+import {FontConfigType} from '../Contexts/SettingsContext';
 
 export const getStyles = (fontConfig: FontConfigType) => {
   const colorRange = {
